+++ src/Pagination/CursorInterface.php
@@ -23,6 +23,7 @@ interface CursorInterface
      *
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function getCurrent();
 
     /**
@@ -30,6 +31,7 @@ interface CursorInterface
      *
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function getPrev();
 
     /**
@@ -37,12 +39,11 @@ interface CursorInterface
      *
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function getNext();
 
     /**
      * Returns the total items in the current cursor.
-     *
-     * @return int
      */
-    public function getCount();
+    public function getCount(): ?int;
 }