+++ src/Resource/NullResource.php
@@ -12,8 +12,6 @@
 namespace League\Fractal\Resource;
 
 /**
- * Null Resource
- *
  * The Null Resource represents a resource that doesn't exist. This can be
  * useful to indicate that a certain relationship is null in some output
  * formats (e.g. JSON API), which require even a relationship that is null at
@@ -26,8 +24,10 @@ class NullResource extends ResourceAbstract
      *
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function getData()
     {
         // Null has no data associated with it.
+        return null;
     }
 }