+++ src/ScopeFactoryInterface.php
@@ -14,26 +14,20 @@ namespace League\Fractal;
 use League\Fractal\Resource\ResourceInterface;
 
 /**
- * ScopeFactoryInterface
- *
  * Creates Scope Instances for resources
  */
 interface ScopeFactoryInterface
 {
-    /**
-     * @param Manager $manager
-     * @param ResourceInterface $resource
-     * @param string|null $scopeIdentifier
-     * @return Scope
-     */
-    public function createScopeFor(Manager $manager, ResourceInterface $resource, $scopeIdentifier = null);
+    public function createScopeFor(
+        Manager $manager,
+        ResourceInterface $resource,
+        ?string $scopeIdentifier = null
+    ): Scope;
 
-    /**
-     * @param Manager $manager
-     * @param Scope $parentScope
-     * @param ResourceInterface $resource
-     * @param string|null $scopeIdentifier
-     * @return Scope
-     */
-    public function createChildScopeFor(Manager $manager, Scope $parentScope, ResourceInterface $resource, $scopeIdentifier = null);
+    public function createChildScopeFor(
+        Manager $manager,
+        Scope $parentScope,
+        ResourceInterface $resource,
+        ?string $scopeIdentifier = null
+    ): Scope;
 }